+++ src/narcoticsense/fusion/__init__.py
@@ -1,0 +1,13 @@
+from narcoticsense.fusion.multimodal import (
+    FusionResult,
+    align_spectral_and_molecular,
+    block_summary,
+    early_fusion_matrix,
+)
+
+__all__ = [
+    "FusionResult",
+    "align_spectral_and_molecular",
+    "block_summary",
+    "early_fusion_matrix",
+]