+++ src/narcoticsense/molecular/__init__.py
@@ -1,0 +1,13 @@
+from narcoticsense.molecular.descriptors import (
+    MoleculeFeatureResult,
+    build_molecular_feature_table,
+    compute_molecular_features,
+    rdkit_available,
+)
+
+__all__ = [
+    "MoleculeFeatureResult",
+    "build_molecular_feature_table",
+    "compute_molecular_features",
+    "rdkit_available",
+]